LayoutTests/platform/mac/accessibility/label-elements-exposed-as-title-ui-elements-expected.txt

 Yes, I'd like TEST .   Yes, I'd like TEST .
This tests that if a radio or checkbox uses a label element, then that label element will be exposed as the title ui element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS checkBoxTitleElement.isEqual(accessibilityController.accessibleElementById('label1')) is true
Checkbox title UI Element role: AXRole: AXGroup
Checkbox title UI Element child count: 2


PASS radioTitleElement.isEqual(accessibilityController.accessibleElementById('label2')) is true
Radio title UI Element role: AXRole: AXGroup
Radio title UI Element child count: 2
PASS successfullyParsed is true

TEST COMPLETE
